# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "mainnet-contract-deployer API Documentation",
        "version": "1.0.0",
        "description": "Moleculer JS Microservice Boilerplate with Typescript, TypeORM, CLI, Service Clients, Swagger, Jest, Docker, Eslint support and everything you will ever need to deploy rock solid projects.."
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
    "paths": {
...
        "/api/v1/request/create": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Request to deploy a contract on mainnet",
                "description": "Request to deploy a contract on mainnet",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "code_ids",
                                    "name",
                                    "email",
                                    "contract_description"
                                ],
                                "properties": {
                                    "code_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        },
                                        "description": "Code ids of all contracts in the group"
                                    },
                                    "name": {
                                        "type": "string",
                                        "description": "Name of subscriber"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "Email of subscriber"
                                    },
                                    "contract_description": {
                                        "type": "string",
                                        "description": "Description of contract(s)"
                                    },
                                    "project_name": {
                                        "type": "string",
                                        "description": "Project name"
                                    },
                                    "official_project_website": {
                                        "type": "string",
                                        "description": "Official project website"
                                    },
                                    "official_project_email": {
                                        "type": "string",
                                        "description": "Official project email"
                                    },
                                    "project_sector": {
                                        "type": "string",
                                        "description": "Sector of the contract in the project"
                                    },
                                    "whitepaper": {
                                        "type": "string",
                                        "description": "Official project whitepaper"
                                    },
                                    "github": {
                                        "type": "string",
                                        "description": "Official project github"
                                    },
                                    "telegram": {
                                        "type": "string",
                                        "description": "Official project telegram"
                                    },
                                    "discord": {
                                        "type": "string",
                                        "description": "Official project discord"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "description": "Official project facebook"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "description": "Official project twitter"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request result"
                    },
                    "422": {
                        "description": "Missing parameters"
                    }
                }
            }
        }
    },
...
}